# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API",
        "description": "api para ver el tema del coso delcoso sobre el cosito(es de comprar cosas)",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "LA api"
        }
    ],
    "paths": {
...
        "/clientes": {
            "get": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Lista a los clientes(hola profe)",
                "operationId": "154539290d4767d4b2fa5730decf1941",
                "responses": {
                    "200": {
                        "description": "se listo todo bien (como esta)"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Agrega un nuevo cliente. (coloque cosas como esta)",
                "operationId": "2554d32d132ab0607fd2b3e89e47911b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "rut_empresa",
                                    "rubro",
                                    "razon_social",
                                    "telefono",
                                    "direccion",
                                    "nombre_contacto",
                                    "email_contacto"
                                ],
                                "properties": {
                                    "rut_empresa": {
                                        "type": "string",
                                        "example": "76123456-7"
                                    },
                                    "rubro": {
                                        "type": "string",
                                        "example": "retail"
                                    },
                                    "razon_social": {
                                        "type": "string",
                                        "example": "Comercial Ejemplo Ltda."
                                    },
                                    "telefono": {
                                        "type": "string",
                                        "example": "+56912345678"
                                    },
                                    "direccion": {
                                        "type": "string",
                                        "example": "Av. Siempre Viva 123"
                                    },
                                    "nombre_contacto": {
                                        "type": "string",
                                        "example": "María López"
                                    },
                                    "email_contacto": {
                                        "type": "string",
                                        "example": "maria@comercialejemplo.cl"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Cliente creado con éxito (pero me quede sin ideas)"
                    },
                    "422": {
                        "description": "Error de validación (y aparte siento que ya repeti mucho)"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/clientes/{cliente}": {
            "get": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Obtiene un cliente ID (yo bien)",
                "operationId": "7d8cc26d0134d372cea97c99fd7ca14d",
                "parameters": [
                    {
                        "name": "cliente",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "obtenido con éxito (la verdad)"
                    },
                    "404": {
                        "description": "no encontrado (en gran parte del codigo)"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Actualiza un cliente por su ID.(una cosa)",
                "operationId": "4a986d8c5549bbe0206255726d15b250",
                "parameters": [
                    {
                        "name": "cliente",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Cliente actualizado con éxito (cuando escribo cosas como)"
                    },
                    "404": {
                        "description": "Cliente no encontrado(eta o pai)"
                    },
                    "422": {
                        "description": "Error de validación(no es falta de ortografia)"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Elimina un cliente por su ID.(asi que)",
                "operationId": "463e28d55aacb1ead7172d869d62c7ba",
                "parameters": [
                    {
                        "name": "cliente",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Cliente eliminado con éxito(como quiero dejarlo solo para este swagger)"
                    },
                    "404": {
                        "description": "Cliente no encontrado(si logra ver esto)"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "el token de segiruda",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}